requirements.txt
asgiref==3.8.1
Django==4.0
django-jazzmin==2.6.0
django-phonenumber-field==7.2.0
django-smtp-ssl==1.0
gunicorn==21.2.0
packaging==24.0
phonenumbers==8.13.34
pillow==10.2.0
psycopg2-binary==2.9.6
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.11.0
